pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stav"
dynamic = ["version"]
description = 'System Trustworthiness and Accountability Vocabulary'
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = []
authors = [{ name = "Arthit Suriyawongkul", email = "suriyawa@tcd.ie" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = []

[project.optional-dependencies]
codegen = ["rdflib>=6.0", "pyyaml>=6.0"]

[project.urls]
Documentation = "https://github.com/unknown/stav#readme"
Issues = "https://github.com/unknown/stav/issues"
Source = "https://github.com/unknown/stav"

[tool.hatch.version]
path = "src/stav/__about__.py"

[tool.hatch.envs.default]
dependencies = ["coverage[toml]>=6.5", "pytest"]

[tool.hatch.envs.codegen]
features = ["codegen"]

[tool.hatch.envs.codegen.scripts]
generate = "python -m stav.codegen.generate {args}"
all = [
  "python -m stav.codegen.generate tools/recipes/dpv_core.json",
  "python -m stav.codegen.generate tools/recipes/dpv_ai.json",
  "python -m stav.codegen.generate tools/recipes/spdx_dataset.json",
  "python -m stav.codegen.generate tools/recipes/spdx_ai.json",
  "python -m stav.codegen.generate tools/recipes/stav_eu_ai.json",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.lint]
detached = true
dependencies = ["black>=23.1.0", "mypy>=1.0.0", "ruff>=0.15.12"]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/stav tests}"
style = ["ruff {args:.}", "black --check --diff {args:.}"]
fmt = ["black {args:.}", "ruff --fix {args:.}", "style"]
all = ["style", "typing"]

[tool.black]
target-version = ["py37"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py37"
line-length = 120
lint.select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105",
  "S106",
  "S107",
  # Ignore complexity
  "C901",
  "PLR0911",
  "PLR0912",
  "PLR0913",
  "PLR0915",
  # Line too long
  "E501",
]
lint.unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.lint.isort]
known-first-party = ["stav"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["stav", "tests"]
branch = true
parallel = true
omit = ["src/stav/__about__.py"]

[tool.coverage.paths]
stav = ["src/stav", "*/stav/src/stav"]
tests = ["tests", "*/stav/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
